neon: Refactor tray, quick actions and push notifications to app level
