+++ packages/neon/lib/src/apps/notifications/app.dart
@@ -1,9 +1,11 @@
 library notifications;
 
 import 'package:flutter/material.dart';
+import 'package:flutter_rx_bloc/flutter_rx_bloc.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:neon/l10n/localizations.dart';
 import 'package:neon/src/apps/notifications/blocs/notifications.dart';
+import 'package:neon/src/blocs/accounts.dart';
 import 'package:neon/src/blocs/apps.dart';
 import 'package:neon/src/neon.dart';
 import 'package:nextcloud/nextcloud.dart';
