+++ packages/neon/lib/src/neon.dart
@@ -9,6 +9,7 @@ import 'package:collection/collection.dart';
 import 'package:crypto/crypto.dart';
 import 'package:file_picker/file_picker.dart';
 import 'package:filesize/filesize.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_cache_manager/flutter_cache_manager.dart';
@@ -33,6 +34,7 @@ import 'package:neon/src/blocs/login.dart';
 import 'package:neon/src/blocs/user_details.dart';
 import 'package:neon/src/blocs/user_status.dart';
 import 'package:neon/src/models/account.dart';
+import 'package:neon/src/models/push_notification_with_account.dart';
 import 'package:nextcloud/nextcloud.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:path/path.dart' as p;
@@ -73,6 +75,7 @@ part 'utils/env.dart';
 part 'utils/global.dart';
 part 'utils/global_options.dart';
 part 'utils/hex_color.dart';
+part 'utils/localizations.dart';
 part 'utils/missing_permission_exception.dart';
 part 'utils/nextcloud_app_specific_options.dart';
 part 'utils/push_utils.dart';
