+++ packages/neon/lib/src/utils/global.dart
@@ -1,7 +1,6 @@
 part of '../neon.dart';
 
 class Global {
-  static Function()? onPushNotificationReceived;
-  static Function(String? payload)? onPushNotificationClicked;
-  static Future<bool> Function(NotificationsNotification notification)? handleNotificationOpening;
+  static Function(String accountID)? onPushNotificationReceived;
+  static Function(PushNotificationWithAccountID notification)? onPushNotificationClicked;
 }
